# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fusion API
  version: 1.0.0
  description: >-
    OpenAPI contract for Fusion backend. All endpoints are under /api.
    Authentication uses a session cookie named `session`.
    Fever compatibility endpoints (/fever, /fever/, /fever.php) are documented
    separately in docs/fever-api.md.
servers:
  - url: /api
...
security:
  - sessionCookie: []
paths:
  /sessions:
    post:
      tags: [Sessions]
      summary: Login with password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
              example:
                data:
                  message: logged in
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Sessions]
      summary: Logout
      security: []
      responses:
        "204":
          description: Logged out
          headers:
            Set-Cookie:
              description: Cleared session cookie
              schema:
                type: string
...
  /oidc/login:
    get:
      tags: [OIDC]
      summary: Get OIDC authorization URL
      description: Endpoint is available only when OIDC is configured.
      security: []
      responses:
        "200":
          description: Authorization URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OIDCLoginEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /oidc/callback:
    get:
      tags: [OIDC]
      summary: OIDC callback endpoint
      description: Exchanges code for identity, creates session, and redirects.
      security: []
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "307":
          description: Redirect to app or login error page
          headers:
            Location:
              schema:
                type: string
  /groups:
    get:
      tags: [Groups]
      summary: List groups
      responses:
        "200":
          description: Group list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Groups]
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpsertRequest"
      responses:
        "200":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /groups/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Groups]
      summary: Get group
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Groups]
      summary: Update group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpsertRequest"
      responses:
        "200":
          description: Group updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Groups]
      summary: Delete group
      description: Group id=1 cannot be deleted. Feeds are moved to group id=1 before delete.
      responses:
        "204":
          description: Group deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /feeds:
    get:
      tags: [Feeds]
      summary: List feeds
      responses:
        "200":
          description: Feed list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Feeds]
      summary: Create feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedRequest"
      responses:
        "200":
          description: Feed created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /feeds/batch:
    post:
      tags: [Feeds]
      summary: Batch create feeds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchCreateFeedsRequest"
      responses:
        "200":
          description: Batch create result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchCreateFeedsEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /feeds/validate:
    post:
      tags: [Feeds]
      summary: Discover feed URLs from input URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateFeedRequest"
      responses:
        "200":
          description: Discovered feeds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateFeedEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /feeds/refresh:
    post:
      tags: [Feeds]
      summary: Start async refresh for all non-suspended feeds
      responses:
        "202":
          description: Refresh job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshJobEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /feeds/refresh-jobs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Feeds]
      summary: Get refresh job status
      responses:
        "200":
          description: Refresh job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshJobEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /feeds/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Feeds]
      summary: Get feed
      responses:
        "200":
          description: Feed detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Feeds]
      summary: Update feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedRequest"
      responses:
        "200":
          description: Feed updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Feeds]
      summary: Delete feed
      description: Deletes feed and its items. Matching saved snapshots keep content with item_id set to null.
      responses:
        "204":
          description: Feed deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /feeds/{id}/refresh:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    post:
      tags: [Feeds]
      summary: Start async refresh for one feed
      responses:
        "202":
          description: Refresh job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshJobEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items:
    get:
      tags: [Items]
      summary: List items
      parameters:
        - name: feed_id
          in: query
          schema:
            type: integer
            format: int64
        - name: group_id
          in: query
          schema:
            type: integer
            format: int64
        - name: unread
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: order_by
          in: query
          schema:
            type: string
            enum: [pub_date, created_at]
            default: pub_date
      responses:
        "200":
          description: Item list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Items]
      summary: Get item
      responses:
        "200":
          description: Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items/-/read:
    patch:
      tags: [Items]
      summary: Mark items as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkItemsRequest"
      responses:
        "204":
          description: Read state updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items/-/unread:
    patch:
      tags: [Items]
      summary: Mark items as unread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkItemsRequest"
      responses:
        "204":
          description: Read state updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /search:
    get:
      tags: [Search]
      summary: Search feeds and items
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Search result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /settings/retention:
    get:
      tags: [Settings]
      summary: Get article retention settings
      responses:
        "200":
          description: Retention settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetentionSettingsEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Settings]
      summary: Update article retention settings and run cleanup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRetentionSettingsRequest"
      responses:
        "200":
          description: Updated retention settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetentionSettingsEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /bookmarks:
    get:
      tags: [Bookmarks]
      summary: List bookmarks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Bookmark list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Bookmarks]
      summary: Create bookmark
      description: >-
        Supports two modes:
        1) provide item_id only; backend snapshots fields from item
        2) provide snapshot fields directly (link, title, content, feed_name; pub_date optional)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/CreateBookmarkFromItemRequest"
                - $ref: "#/components/schemas/CreateBookmarkSnapshotRequest"
      responses:
        "200":
          description: Bookmark created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /bookmarks/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Bookmarks]
      summary: Get bookmark
      responses:
        "200":
          description: Bookmark detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Bookmarks]
      summary: Delete bookmark
      responses:
        "204":
          description: Bookmark deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /read-later:
    get:
      tags: [ReadLater]
      summary: List read later items
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          description: Opaque cursor returned as next_cursor from the previous page.
          schema:
            type: string
      responses:
        "200":
          description: Read later item list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadLaterItemListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [ReadLater]
      summary: Create read later item
      description: >-
        Supports two modes:
        1) provide item_id only; backend snapshots fields from item
        2) provide snapshot fields directly (link, title, content, feed_name; pub_date optional)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/CreateReadLaterItemFromItemRequest"
                - $ref: "#/components/schemas/CreateReadLaterItemSnapshotRequest"
      responses:
        "200":
          description: Read later item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadLaterItemEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /read-later/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [ReadLater]
      summary: Get read later item
      responses:
        "200":
          description: Read later item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadLaterItemEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [ReadLater]
      summary: Delete read later item
      responses:
        "204":
          description: Read later item deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: invalid request
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: unauthorized
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: resource not found
    TooManyRequests:
      description: Too many login attempts
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: too many login attempts
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: internal server error
  schemas:
...
    MessageEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Message"
    LoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
    GroupUpsertRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
...
    GroupEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Group"
    GroupListEnvelope:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        total:
          type: integer
...
    FeedEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Feed"
    FeedListEnvelope:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Feed"
        total:
          type: integer
...
    ValidateFeedRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    OIDCLoginEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/OIDCLoginData"